pyproject.toml
[project]
name = "trellis"
version = "0.1.0"
description = "TRELLIS - Scalable and Versatile 3D Generation from images"
readme = "README.md"
requires-python = ">=3.10,<3.11"
license = "MIT"
dependencies = [
    "torch==2.4.0+cu121",
    "torchvision==0.19.0+cu121",
    "pillow==10.4.0",
    "imageio==2.36.1",
    "imageio-ffmpeg==0.5.1",
    "tqdm==4.67.1",
    "easydict==1.13",
    "opencv-python-headless==4.10.0.84",
    "scipy==1.14.1",
    "rembg==2.0.60",
    "onnxruntime==1.20.1",
    "trimesh==4.5.3",
    "xatlas==0.0.9",
    "pyvista==0.44.2",
    "pymeshfix==0.17.0",
    "igraph==0.11.8",
    "utils3d @ git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8",
    "xformers==0.0.27.post2",
    "spconv-cu120==2.3.6",
    "transformers==4.46.3",
    "gradio>=4.0,<5.0",
    "gradio_litmodel3d==0.0.1",
    "pydantic==2.10.6",
    "spaces>=0.44.0",
    "diff-gaussian-rasterization @ git+https://github.com/ashawkey/diff-gaussian-rasterization.git",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["trellis"]

[tool.uv]
# Use CUDA 12.1 PyTorch wheels as a supplementary index
index-url = "https://pypi.org/simple"
extra-index-url = ["https://download.pytorch.org/whl/cu121"]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
# Git dependency for utils3d
utils3d = { git = "https://github.com/EasternJournalist/utils3d.git", rev = "9a4eb15e4021b67b12c460c7057d642626897ec8" }

[project.scripts]
trellis = "app:main"
